requirements.txt
streamlit
wikipedia
deep_translator
python-docx
requests
